chore: create deprecated directory in shared & change filename
