+++ libs/shared/deprecated/apis/githubApi/githubApiCommon.api.ts
@@ -1,15 +1,15 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 import { AxiosRequestConfig, InternalAxiosRequestConfig } from 'axios'
 
-import { CommonResponse } from '../../types/instance.type'
-import { githubRawInstance } from './githubRawInstance.api'
+import { CommonResponse } from '../../../types/instance.type'
+import { githubApiInstance } from './githubApiInstance.api'
 
 /* get 요청 */
-export const getRequestGithubRaw = async <T>(
+export const getRequestGithubApi = async <T>(
   url: string,
   config?: AxiosRequestConfig,
 ): Promise<T> => {
-  const response = await githubRawInstance.get<CommonResponse<T>>(
+  const response = await githubApiInstance.get<CommonResponse<T>>(
     url,
     config as InternalAxiosRequestConfig,
   )
@@ -17,12 +17,12 @@ export const getRequestGithubRaw = async <T>(
 }
 
 // /* post 요청 */
-export const postRequestGithubRaw = async <T>(
+export const postRequestGithubApi = async <T>(
   url: string,
   data?: any,
   config?: AxiosRequestConfig,
 ): Promise<T> => {
-  const response = await githubRawInstance.post<CommonResponse<T>>(
+  const response = await githubApiInstance.post<CommonResponse<T>>(
     url,
     data,
     config as InternalAxiosRequestConfig,
@@ -31,11 +31,11 @@ export const postRequestGithubRaw = async <T>(
 }
 
 /* delete 요청 */
-export const deleteRequestGithubRaw = async <T>(
+export const deleteRequestGithubApi = async <T>(
   url: string,
   config?: AxiosRequestConfig,
 ): Promise<T> => {
-  const response = await githubRawInstance.delete<CommonResponse<T>>(
+  const response = await githubApiInstance.delete<CommonResponse<T>>(
     url,
     config as InternalAxiosRequestConfig,
   )
@@ -43,12 +43,12 @@ export const deleteRequestGithubRaw = async <T>(
 }
 
 /* put 요청 */
-export const putRequestGithubRaw = async <T>(
+export const putRequestGithubApi = async <T>(
   url: string,
   data: any,
   config?: AxiosRequestConfig,
 ): Promise<T> => {
-  const response = await githubRawInstance.put<CommonResponse<T>>(
+  const response = await githubApiInstance.put<CommonResponse<T>>(
     url,
     data,
     config as InternalAxiosRequestConfig,
@@ -57,12 +57,12 @@ export const putRequestGithubRaw = async <T>(
 }
 
 /* patch 요청 */
-export const patchRequestGithubRaw = async <T>(
+export const patchRequestGithubApi = async <T>(
   url: string,
   data: any,
   config?: AxiosRequestConfig,
 ): Promise<T> => {
-  const response = await githubRawInstance.patch<CommonResponse<T>>(
+  const response = await githubApiInstance.patch<CommonResponse<T>>(
     url,
     data,
     config as InternalAxiosRequestConfig,
